Move AccountSettings to components directory
