+++ src/pages/views/overview/overviewWrapper.tsx
@@ -1,4 +1,4 @@
-import { AccountSettings } from 'pages/views/components/accountSettings';
+import { AccountSettings } from 'components/accountSettings';
 import React from 'react';
 import { RouteComponentProps } from 'react-router-dom';
